Super refactor

+++ Backend/src/main/java/fpt/edu/vn/Backend/DTO/response/CurrencyExchangeRateResponse.java
@@ -5,11 +5,12 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 @Builder
 public class CurrencyExchangeRateResponse implements Serializable {
-    private Map<CurrencyType, Double> rates;
+    private Map<CurrencyType, BigDecimal> rates;
     private long lastUpdate;
 }
